+++ R/fit-action-model.R
@@ -14,7 +14,7 @@
 #'
 #' @includeRmd man/rmd/indicators.Rmd details
 #'
-#' @inheritParams ellipsis::dots_empty
+#' @inheritParams rlang::args_dots_empty
 #'
 #' @param x A workflow.
 #'
@@ -60,7 +60,7 @@
 #' update_model(fitted, regularized_model)
 #'
 add_model <- function(x, spec, ..., formula = NULL) {
-  ellipsis::check_dots_empty()
+  check_dots_empty()
   action <- new_action_model(spec, formula)
   add_action(x, action, "model")
 }
@@ -86,7 +86,7 @@ remove_model <- function(x) {
 #' @rdname add_model
 #' @export
 update_model <- function(x, spec, ..., formula = NULL) {
-  ellipsis::check_dots_empty()
+  check_dots_empty()
   x <- remove_model(x)
   add_model(x, spec, formula = formula)
 }